Added following methods on assessor controller:     1. assessor_auth     2. get_by_wallet     3. get_by_wallet_with_provider
Added all the required assessor router paths associated with the controller methods added

Updated some readme data on some of the assessor router paths that were added previously.

Added primary generated column on Assessor entity due to need of relation between the complain>

Removed sources enum as the consent dates for assessor and provider are now held separately.

Updated some methods used in the complaint router that previously reffered to provider and now>

Fixed queries to reflect the separation of the assessor-provider entities.